+++ src/extension.ts
@@ -8,6 +8,7 @@ import { registerXmlPreview } from "./xmlpreview";
 import { registerQuickFix } from "./quickfix";
 import { registerLanguageFeatures } from "./language";
 import { registerCodeLens } from "./codelens";
+import { registerViewPreview } from "./viewpreview";
 import { classNameOf, isAppClass } from "./abap";
 import { originOf } from "./urls";
 import { registerAppSearch } from "./appsearch";
@@ -339,6 +340,7 @@ export function activate(context: vscode.ExtensionContext): void {
   registerPropertyEditor(context, log);
   registerViewCheck(context, log);
   registerXmlPreview(context, log, findingsNow);
+  registerViewPreview(context, log);
   registerQuickFix(context, log);
   registerLanguageFeatures(context, log);
   registerCodeLens(context);